pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "threesa"
version = "0.1.0"
description = "Research scaffold for the 3SA LVLM adversarial attack."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "pillow>=10.0",
  "scikit-image>=0.22",
  "torch>=2.2",
  "torchvision>=0.17",
  "transformers>=4.40",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
